backend/requirements.txt
flask==3.1.1
flask-cors==5.0.1
python-dotenv==1.1.0
python-docx==1.1.2
PyMuPDF==1.25.5
groq==0.25.0
werkzeug==3.1.3
langdetect==1.0.9
pygments==2.19.1
nltk==3.9.1
requests==2.32.3
